split types into interface files
