+++ src/components/GameLetters/GameLetters.test.tsx
@@ -4,7 +4,7 @@ import { render, screen } from '@testing-library/react';
 import { playText } from '../../utils/constants.util';
 import { levelWordsData, gameLettersData } from '../../utils/tests.util';
 import GameLetters from './GameLetters';
-import { Letter } from '../../types/types';
+import { Letter } from '../../types/letter.interface';
 
 describe('Register component', () => {
   const gameLetters = gameLettersData() as Letter[];
