+++ src/types/letter.interface.ts
@@ -1,0 +1,6 @@
+export interface Letter {
+  id: number,
+  char: string,
+  updatedAt: number,
+  used: boolean
+}
