+++ src/types/notification.interface.ts
@@ -1,0 +1,7 @@
+import { notifications } from '../utils/constants.util';
+
+export type NotificationKey = keyof typeof notifications;
+
+export type NotificationProps = {
+  name: NotificationKey;
+};
