+++ src/types/start-page.interface.ts
@@ -1,0 +1,6 @@
+import { GameStatus } from './game.interface';
+
+export interface StartPageProps {
+  updateGameStatus: (status: GameStatus) => void,
+  hasLevelWord: boolean
+}
